requirements.txt
streamlit
python-dotenv
openai
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
chromadb